# [cpp] Optimize docker pull by using docker compose