--- a/.github/actions/pull_images/action.yml
+++ b/.github/actions/pull_images/action.yml
@@ -30,6 +30,8 @@ runs:
       run: |
         source venv/bin/activate
         python utils/scripts/get-image-list.py '${{ inputs.scenarios }}' > compose.yaml
+      env:
+        PYTHONPATH: "."
 
     - name: Pull
       shell: bash