## Motivation

Before :
* simple run (only default scenario) : 12s https://github.com/DataDog/system-tests/actions/runs/9680233476/job/26708097007#step:4:1
* full run (all scenario) : 219s https://github.com/DataDog/system-tests/actions/runs/9680330556/job/26713339712#step:4:1

After : 

* simple run (only default scenario) : 13s https://github.com/DataDog/system-tests/actions/runs/9684405073/job/26722150197?pr=2611
* full run (all scenarios): 97s https://github.com/DataDog/system-tests/actions/runs/9684470124/job/26722530700#step:4:1

These figures are collected the exact same list.

Turns out we misses two images : 


* localstack/localstack:3.1.0                   1.09GB   CROSSED_TRACING_LIBRARIES
* otel/opentelemetry-collector-contrib:latest   255MB (OpenTelemetryScenarios)

With those two images, we add 30s the the pulling, BUT it 's saved during the scenario execution


## Changes

* [x] dynamically generate the list of image to pull
* [x] use `docker compose pull` to benefit from parallel pulling

## Workflow


1. ⚠️ Create your PR as draft ⚠️
2. Work on you PR until the CI passes (if something not related to your task is failing, you can ignore it)
3. Mark it as ready for review
    * Test logic is modified? -> Get a review from RFC owner. We're working on refining the `codeowners` file quickly.
    * Framework is modified, or non obvious usage of it -> get a review from [R&P team](https://dd.enterprise.slack.com/archives/C025TJ4RZ8X)

:rocket: Once your PR is reviewed, you can merge it!

🛟 [#apm-shared-testing](https://dd.enterprise.slack.com/archives/C025TJ4RZ8X) 🛟

## Reviewer checklist

* [ ] [Relevant label](https://github.com/DataDog/system-tests/blob/main/docs/CI/labels.md) (`run-parametric-scenario`, `run-profiling-scenario`...) are presents
* [ ] If PR title starts with `[<language>]`, double-check that only `<language>` is impacted by the change
* [ ] No system-tests internal is modified. Otherwise, I have the approval from [R&P team](https://dd.enterprise.slack.com/archives/C025TJ4RZ8X)
* [ ] CI is green, or failing jobs are not related to this change (and you are 100% sure about this statement)
* [ ] A docker base image is modified?
    * [ ] the relevant `build-XXX-image` label is present
* [ ] A scenario is added (or removed)?
    * [ ] Get a review from [R&P team](https://dd.enterprise.slack.com/archives/C025TJ4RZ8X)